Clean up and add error message
