+++ projman_filler/lane_level_statistics.py
@@ -1,6 +1,8 @@
 
 from collections import defaultdict
 import math
+import sys
+
 from projman_filler.models.db_models import FlowcellLaneResult
 
 
@@ -38,11 +40,19 @@ def calculate_lane_statistics(flowcell_name, conversion_results, reads_and_cycle
 
         for read_nbr in reads_and_cycles.keys():
             cycles = reads_and_cycles[read_nbr]
-
-            if math.isnan(error_rates[lane_nbr][read_nbr]):
-                error_rate = None
-            else:
-                error_rate = error_rates[lane_nbr][read_nbr]
+            try:
+                if math.isnan(error_rates[lane_nbr][read_nbr]):
+                    error_rate = None
+                else:
+                    error_rate = error_rates[lane_nbr][read_nbr]
+            except KeyError as e:
+                key_nbr = e.args[0]
+                if key_nbr == 3:
+                    print("Found a third read in the in the Stats.json file, but no third read in the InterOp data, "
+                          "perhaps you want to use --atac-seq-mode?")
+                    sys.exit(1)
+                else:
+                    raise e
 
             raw_density = densities[lane_nbr][read_nbr]["raw_density"]
             pf_density = densities[lane_nbr][read_nbr]["pass_filter_density"]
